+++ frontend/src/pages/games/Games.tsx
@@ -4,7 +4,6 @@ import GameCard from '@/components/games/GameCard'
 import { gamesApi, housesApi } from '@/lib/api'
 import { useRtpSocket } from '@/hooks/useRtpSocket'
 import { Game, BettingHouse } from '@/types'
-import { applySignedInt } from '@/lib/utils'
 import Input from '@/components/ui/Input'
 import Button from '@/components/ui/Button'
 
@@ -83,13 +82,11 @@ export default function GamesPage() {
     period: 'daily' | 'weekly' | 'monthly'
   ): number => {
     const up = updates.find((u) => u.gameName === game.name && u.houseId === houseId)
-    const base = applySignedInt(game.rtpDecimal ?? 0, game.signedInt)
     let raw: number | undefined
     if (period === 'daily') raw = up?.rtpDaily
     else if (period === 'weekly') raw = up?.rtpWeekly
     else raw = up?.rtpMonthly
-    const rawRtp = raw ?? base
-    return rawRtp / 100
+    return (raw ?? 0) / 100
   }
 
   const toggleExpanded = (houseId: number) => {
